+++ resolwe/flow/executors/__main__.py
@@ -29,6 +29,7 @@ using the python from the ``venv`` virtualenv.
 import argparse
 import asyncio
 import logging
+import os
 import sys
 import traceback
 from contextlib import suppress
@@ -44,6 +45,11 @@ from .socket_utils import Response, ResponseStatus
 from .zeromq_utils import ZMQCommunicator
 
 logger = logging.getLogger(__name__)
+
+# Secrets necessary to connect to the listener service.
+LISTENER_PUBLIC_KEY = os.getenv("LISTENER_PUBLIC_KEY").encode()
+PUBLIC_KEY = os.getenv("CURVE_PUBLIC_KEY").encode()
+PRIVATE_KEY = os.getenv("CURVE_PRIVATE_KEY").encode()
 
 
 def handle_exception(exc_type, exc_value, exc_traceback):
@@ -70,6 +76,9 @@ async def open_listener_connection(data_id, host, port, protocol) -> ZMQCommunic
     """Connect to the listener service."""
     zmq_context = zmq.asyncio.Context.instance()
     zmq_socket = zmq_context.socket(zmq.DEALER)
+    zmq_socket.curve_secretkey = PRIVATE_KEY
+    zmq_socket.curve_publickey = PUBLIC_KEY
+    zmq_socket.curve_serverkey = LISTENER_PUBLIC_KEY
     zmq_socket.setsockopt(zmq.IDENTITY, f"-{data_id}".encode())
     connect_string = f"{protocol}://{host}:{port}"
     zmq_socket.connect(connect_string)