+++ docs/CHANGELOG.rst
@@ -15,6 +15,8 @@ Changed
 - **BACKWARD INCOMPATIBLE:** Make move between collections a background job
 - **BACKWARD INCOMPATIBLE:** Remove support for ``Python 3.10``
 - Use ``simple_unaccent`` full text search configuration instead of ``simple``
+- Authenticate worker with the listener (and vice versa) and encrypt the 
+  communication between them using ``CurveZMQ`` protocol
 
 
 ===================